Several changes: -- Tweaks to the cutoffs for expansions (log_gev_pdf had a typo with 1e-4) -- Change box constraints on parameters to be closer to zero (primarily a debugging experiment for tests) -- Remove RCall dep in tests, add Serialization dep -- Add serialized R mle results from ismev for testing -- Update the JuMP for specifying Ipopt, per a warning message
The current testing fails several times, however, for one specific
distribution: GEV(1, 1, 1/2). And the failure is JuMP or Ipopt related.
All estimates that return with a success code pass the test comparing
with the R library ismev. So the only thing left to do is figure out the
GEV(1,1,1/2) failure issue.